refactor(addons): Use `index.ts` as main
